remove path from default values for kibana ingress
the default helm chart values for kibana ingress contains `kibana.ingress.path: /`

this value is non-functional and misleading.  the path is derived from the entries in kibana.ingress.hosts if they have an additional path beyond the domain.

this change removes `kibana.ingress.path` and provides a commented example in the values.yaml file to demonstrate how paths can be set correctly.