Refactor and use es_sink for BGG and iot_device_simulator
I've refactored classes in the es_sink library to make it easier to use with basic auth and with AWS auth. I refactored BGG to use the new es_sink calls. I removed the same classes from iot_device_simulator and added es_sink as a dependency. Also fixes a bug in BGG, where it did not refer to self.